Restructure demo according to platform-agnostic protocol (#4)
The restructuring is grand in the sense that it touches every part of the app. The purpose is to introduce the use of JSON-LD and RO-Crate in a manner closer to the interoperability protocol and standards discussed and developed at the MADICES workshop. A demonstration of the protocol is available on the PREMISE website's tutorials page.